Add comments and fix prototype to match type naming convention.

+++ src/util.h
@@ -2,6 +2,16 @@
 #define __util_h
 #include <cat/cattypes.h>
 
+/* 
+ * Compute the 1s complement sum of 'len' bytes.  'val' holds the
+ * checksum thusfar.  (usually starts at 0).  This function 
+ * assumes that the checksum starts on a 2-byte aligned boundary.
+ * Because a 1s complement sum is endian neutral, one can compute
+ * the checksum and then swap the bytes of the result as a
+ * post-processing step and get the same result as if the entire sum
+ * were computed by first performing appropriate byte swapping when
+ * reading the 16-bit values.  
+ */
 uint16_t ones_sum(void *p, size_t len, uint16_t val);
 
 /*
@@ -19,8 +29,8 @@ uint16_t ones_sum(void *p, size_t len, uint16_t val);
  *   uint32_t flowlabel = bitfield((byte_t *)v6p, 12, 20);
  *   
  */
-unsigned long getbitfield(const byte_t * p, size_t bitoff, size_t bitlen);
-void setbitfield(byte_t * p, size_t bitoff, size_t bitlen, unsigned long val);
+ulong getbitfield(const byte_t * p, size_t bitoff, size_t bitlen);
+void setbitfield(byte_t * p, size_t bitoff, size_t bitlen, size_t val);
 
 /* return 1 if bit #n is set in p */
 int getbit(const byte_t * p, size_t n);
